re-organize existing dataset code to prepare ibis relation integration
